Add exhaustion.sync.threshold config option to control the frequency of exhaustion sync packets

+++ java/squeek/applecore/ModConfig.java
@@ -17,8 +17,8 @@ public class ModConfig
 	private static final String CATEGORY_SERVER_COMMENT =
 		"These config settings are server-side only";
 
-	public static float EXHAUSTION_SYNC_THRESHOLD = ModConfig.EXHAUSTION_SYNC_THRESHOLD_DEFAULT;
-	public static float EXHAUSTION_SYNC_THRESHOLD_DEFAULT = 0.01f;
+	public static double EXHAUSTION_SYNC_THRESHOLD = ModConfig.EXHAUSTION_SYNC_THRESHOLD_DEFAULT;
+	public static double EXHAUSTION_SYNC_THRESHOLD_DEFAULT = 0.01D;
 	private static final String EXHAUSTION_SYNC_THRESHOLD_NAME = "exhaustion.sync.threshold";
 	private static final String EXHAUSTION_SYNC_THRESHOLD_COMMENT =
 		"The maximum difference between the server's value for exhaustion and the client's before the value is syncronized from the server to the client.\n"
@@ -88,7 +88,7 @@ public class ModConfig
 		 */
 		config.getCategory(CATEGORY_SERVER).setComment(CATEGORY_SERVER_COMMENT);
 
-		EXHAUSTION_SYNC_THRESHOLD = (float) config.get(CATEGORY_SERVER, EXHAUSTION_SYNC_THRESHOLD_NAME, EXHAUSTION_SYNC_THRESHOLD_DEFAULT, EXHAUSTION_SYNC_THRESHOLD_COMMENT).getDouble(EXHAUSTION_SYNC_THRESHOLD_DEFAULT);
+		EXHAUSTION_SYNC_THRESHOLD = config.get(CATEGORY_SERVER, EXHAUSTION_SYNC_THRESHOLD_NAME, EXHAUSTION_SYNC_THRESHOLD_DEFAULT, EXHAUSTION_SYNC_THRESHOLD_COMMENT).getDouble(EXHAUSTION_SYNC_THRESHOLD_DEFAULT);
 
 		/*
 		 * CLIENT
